apps/api/pyproject.toml
[project]
name = "heme-kiitchen-api"
version = "0.1.0"
description = "Hémé Kiitchen FastAPI backend"
authors = [
    {name = "Bhavyesh Shah"}
]
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standard] (>=0.43.0,<0.44.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
